fix index issue and update auth module
