+++ auth/middleware.go
@@ -67,30 +67,6 @@ func (au *AuthHandler) IsAuthenticated(nextHandler http.HandlerFunc) http.Handle
 	}
 }
 
-// OptionalAuthenticated calls the next's handler's ServeHTTP() with the request context unchanged
-// if a user is not authenticated, else it modifies the request context with a copy of the user's
-// details and passes the changed copy of the request to the next handler's ServeHTTP().
-func (au *AuthHandler) OptionalAuthentication(nextHandler http.HandlerFunc) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Add("content-type", "application/json")
-
-		store := NewMongoStore(utils.GetCollection(sessionCollection), au.configs.SessionMaxAge, true, []byte(au.configs.SecretKey))
-		_, er := store.Get(r, au.configs.SessionKey)
-		status, sessData, err := GetSessionDataFromToken(r, []byte(au.configs.HmacSampleSecret))
-
-		if er != nil || err != nil {
-			if !status && sessData.Email == "" {
-				nextHandler.ServeHTTP(w, r)
-				return
-			}
-		}
-
-		ctx := context.WithValue(r.Context(), UserDetails, &sessData)
-		r = r.WithContext(ctx)
-		nextHandler.ServeHTTP(w, r)
-	}
-}
-
 func (au *AuthHandler) IsAuthorized(nextHandler http.HandlerFunc, role string) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json")
@@ -157,5 +133,29 @@ func (au *AuthHandler) IsAuthorized(nextHandler http.HandlerFunc, role string) h
 		//nolint:staticcheck //CODEI8: lint ignore
 		ctx := context.WithValue(r.Context(), UserContext, u)
 		nextHandler.ServeHTTP(w, r.WithContext(ctx))
+	}
+}
+
+// OptionalAuthenticated calls the next's handler's ServeHTTP() with the request context unchanged
+// if a user is not authenticated, else it modifies the request context with a copy of the user's
+// details and passes the changed copy of the request to the next handler's ServeHTTP().
+func (au *AuthHandler) OptionalAuthentication(nextHandler http.HandlerFunc) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Add("content-type", "application/json")
+
+		store := NewMongoStore(utils.GetCollection(sessionCollection), au.configs.SessionMaxAge, true, []byte(au.configs.SecretKey))
+		_, er := store.Get(r, au.configs.SessionKey)
+		status, sessData, err := GetSessionDataFromToken(r, []byte(au.configs.HmacSampleSecret))
+
+		if er != nil || err != nil {
+			if !status && sessData.Email == "" {
+				nextHandler.ServeHTTP(w, r)
+				return
+			}
+		}
+
+		ctx := context.WithValue(r.Context(), UserDetails, &sessData)
+		r = r.WithContext(ctx)
+		nextHandler.ServeHTTP(w, r)
 	}
 }
