+++ organizations/payments.go
@@ -1,14 +1,8 @@
 package organizations
 
 import (
-	"crypto/aes"
-	"crypto/cipher"
-	"crypto/md5"
-	"crypto/rand"
-	"encoding/hex"
 	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"os"
 	"strconv"
@@ -44,31 +38,6 @@ const (
 	Visa            = "Visa"
 	UnknownCard     = "Unknown"
 )
-
-func createHash(key string) string {
-	hasher := md5.New()
-	hasher.Write([]byte(key))
-	
-	return hex.EncodeToString(hasher.Sum(nil))
-}
-
-func encrypt(data []byte, passphrase string) []byte {
-	block, _ := aes.NewCipher([]byte(createHash(passphrase)))
-	gcm, err := cipher.NewGCM(block)
-	
-	if err != nil {
-		panic(err.Error())
-	}
-	
-	nonce := make([]byte, gcm.NonceSize())
-	if _, err = io.ReadFull(rand.Reader, nonce); err != nil {
-		panic(err.Error())
-	}
-	
-	ciphertext := gcm.Seal(nonce, nonce, data, nil)
-
-	return ciphertext
-}
 
 // Converts amount in real currency to equivalent token value.
 func GetTokenAmount(amount float64, currency string) (float64, error) {
@@ -405,8 +374,8 @@ func (oh *OrganizationHandler) AddCard(w http.ResponseWriter, r *http.Request) {
 
 	newcard.OrgID = orgID
 	newcard.MemberID = MemberID
-	newcard.CVCCheck = string(encrypt([]byte(newcard.CVCCheck), "zcore_key"))
-	newcard.CardNumber = string(encrypt([]byte(newcard.CardNumber), "zcore_key"))
+	newcard.CVCCheck = string(utils.GCMEncrypt([]byte(newcard.CVCCheck), "zcore_key"))
+	newcard.CardNumber = string(utils.GCMEncrypt([]byte(newcard.CardNumber), "zcore_key"))
 
 	// convert card struct to map
 	card, err := utils.StructToMap(newcard)
